requirements.txt
# Core Web & Database
Flask
Flask-Login
gunicorn
uvicorn
SQLAlchemy
pymysql
python-dotenv
dramatiq[redis]
redis 

# Data Handling & APIs
pandas
requests # For calling Ollama API
aiohttp
PyYAML
PyJWT
Jinja2
asgiref